+++ app/src/main/java/com/shinjiindustrial/portmapper/domain/AndroidUpnpServiceConfiguration.kt
@@ -1,0 +1,54 @@
+package com.shinjiindustrial.portmapper.domain
+
+import android.content.Context
+import com.shinjiindustrial.portmapper.NetworkType
+import org.fourthline.cling.android.AndroidNetworkAddressFactory
+import org.fourthline.cling.android.AndroidUpnpServiceConfiguration
+import org.fourthline.cling.binding.xml.ServiceDescriptorBinder
+import org.fourthline.cling.binding.xml.UDA10ServiceDescriptorBinderImpl
+import org.fourthline.cling.transport.spi.NetworkAddressFactory
+
+class AndroidUpnpServiceConfigurationImpl(context : Context) : AndroidUpnpServiceConfiguration() {
+
+    var Context : Context
+    init {
+        Context = context
+    }
+
+    override fun getServiceDescriptorBinderUDA10(): ServiceDescriptorBinder {
+        return UDA10ServiceDescriptorBinderImpl()
+    }
+
+    // in case you want to get additional info from initialization
+    // we can then maybe check if they support multicast (i.e. data typically does not)
+    // and throw away.  also we can use connectionmanager to enumerate interfaces
+    // by default this returns new AndroidNetworkAddressFactory
+    // AndroidNetworkAddressFactory has a method discoverNetworkInterfaces() which
+    //   calls the java NetworkInterface.getNetworkInterfaces();
+
+    override fun createNetworkAddressFactory(streamListenPort: Int): NetworkAddressFactory? {
+        //return NetworkAddressFactoryImpl(streamListenPort)
+        val addressFactory = AndroidNetworkAddressFactory(streamListenPort)
+
+        // the final set of usable interfaces and bind addresses
+        val iterator = addressFactory.networkInterfaces
+        val networkInterfaces: MutableList<java.net.NetworkInterface> = ArrayList()
+        while (iterator.hasNext()) {
+            networkInterfaces.add(iterator.next())
+        }
+        NetworkInterfacesUsed = networkInterfaces
+        NetworkMappings = OurNetworkInfo.GetNameTypeMappings(Context)
+
+        NetworkInterfacesUsedInfos = mutableListOf()
+        for (netInterface in networkInterfaces)
+        {
+            NetworkInterfacesUsedInfos?.add(Pair<java.net.NetworkInterface, NetworkType>(netInterface, OurNetworkInfo.GetTypeFromInterfaceName(NetworkMappings, netInterface.name)))
+        }
+
+        return addressFactory
+    }
+
+    var NetworkInterfacesUsed : MutableList<java.net.NetworkInterface>? = null
+    var NetworkMappings : MutableMap<String, NetworkType>? = null
+    var NetworkInterfacesUsedInfos : MutableList<Pair<java.net.NetworkInterface, NetworkType>>? = null
+}
